refactor(api): extract last 5 sub-routers from system.rs (#3749 11/N)
Closes #3749.

Drains the last set of handlers from system.rs and lands the issue's
acceptance: each sub-router under 300 lines, system.rs reduced to a
thin composer + the workspace-level /api/versions metadata + the two
cross-crate path helpers (librefang_home, hostname_string) that
sibling routers still call into.

Final extractions:

- New routes/commands.rs (115 lines)
  list_commands, get_command + /commands{,/:name} routes.
  Fixed builtin slash-command catalog plus skill-derived dynamic entries.

- New routes/bindings.rs (55 lines)
  list_bindings, add_binding, remove_binding + /bindings{,/:index}.

- New routes/logs.rs (115 lines)
  logs_stream + classify_audit_level helper + /logs/stream SSE.

- Extended routes/webhooks.rs
  webhook_wake, webhook_agent + validate_webhook_token,
  webhook_unauthorized_response helpers + /hooks/{wake,agent} routes.
  Module-doc rewritten so the tri-flavor scope (subscriptions, outbound,
  external triggers) is explicit.

- Extended routes/memory.rs
  6 KV handlers (get_agent_kv, get_agent_kv_key, set_agent_kv_key,
  delete_agent_kv_key, export_agent_memory, import_agent_memory) +
  /memory/agents/{id}/kv* and /agents/{id}/memory/{export,import}
  routes. Imports promoted (AgentId, ErrorTranslator, AgentIdPath,
  RequestLanguage) so handlers don't carry full paths.

- Extended routes/task_queue.rs
  queue_status + /queue/status route.

- Rewritten routes/system.rs (965 → 117 lines)
  Now contains only: composing router that .merge()s the extracted
  per-domain routers, /api/versions metadata handler, and the
  librefang_home / hostname_string path helpers (still pub(super) /
  pub(crate) — agent_templates and backup call into them).

- routes/mod.rs registers the three new public modules
  (bindings, commands, logs) and re-exports them via the
  back-compat `pub use` glob.

- tests/hooks_commands_routes_integration.rs harness updated
  Mounts webhooks::router() alongside system::router() so /api/hooks/*
  is reachable; system::router() already merges commands::router()
  per the new layout, so commands routes don't need a second merge
  (axum panics on overlapping method routes).

Verification:
- cargo check -p librefang-api --lib  → clean
- cargo clippy -p librefang-api --all-targets -- -D warnings  → clean
- cargo test -p librefang-api integration suites
  - hooks_commands_routes_integration  → 16/16
  - memory_routes_integration          → 12/12
  - system_tools_sessions_test         → 10/10
  - api_integration_test               → 72/72
  - channel_webhooks_test              → 4/4
  Total: 114 integration tests passing.
