+++ crates/librefang-api/src/routes/logs.rs
@@ -1,0 +1,121 @@
+//! Real-time audit log streaming (SSE) endpoint (#3749 11/N).
+
+use super::AppState;
+use axum::extract::{Query, State};
+use axum::response::IntoResponse;
+use std::collections::HashMap;
+use std::sync::Arc;
+
+pub fn router() -> axum::Router<Arc<AppState>> {
+    axum::Router::new().route("/logs/stream", axum::routing::get(logs_stream))
+}
+
+/// GET /api/logs/stream — SSE endpoint for real-time audit log streaming.
+///
+/// Streams new audit entries as Server-Sent Events. Accepts optional query
+/// parameters for filtering:
+///   - `level`  — filter by classified level (info, warn, error)
+///   - `filter` — text substring filter across action/detail/agent_id
+///   - `token`  — auth token (for EventSource clients that cannot set headers)
+///
+/// A heartbeat ping is sent every 15 seconds to keep the connection alive.
+/// The endpoint polls the audit log every second and sends only new entries
+/// (tracked by sequence number). On first connect, existing entries are sent
+/// as a backfill so the client has immediate context.
+#[utoipa::path(get, path = "/api/logs/stream", tag = "system", responses((status = 200, description = "SSE log stream")))]
+pub async fn logs_stream(
+    State(state): State<Arc<AppState>>,
+    Query(params): Query<HashMap<String, String>>,
+) -> axum::response::Response {
+    use axum::response::sse::{Event, KeepAlive, Sse};
+
+    let level_filter = params.get("level").cloned().unwrap_or_default();
+    let text_filter = params
+        .get("filter")
+        .cloned()
+        .unwrap_or_default()
+        .to_lowercase();
+
+    let (tx, rx) = tokio::sync::mpsc::channel::<
+        Result<axum::response::sse::Event, std::convert::Infallible>,
+    >(256);
+
+    tokio::spawn(async move {
+        let mut last_seq: u64 = 0;
+        let mut first_poll = true;
+
+        loop {
+            tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+
+            let entries = state.kernel.audit().recent(200);
+
+            for entry in &entries {
+                // On first poll, send all existing entries as backfill.
+                // After that, only send entries newer than last_seq.
+                if !first_poll && entry.seq <= last_seq {
+                    continue;
+                }
+
+                let action_str = format!("{:?}", entry.action);
+
+                // Apply level filter
+                if !level_filter.is_empty() {
+                    let classified = classify_audit_level(&action_str);
+                    if classified != level_filter {
+                        continue;
+                    }
+                }
+
+                // Apply text filter
+                if !text_filter.is_empty() {
+                    let haystack = format!("{} {} {}", action_str, entry.detail, entry.agent_id)
+                        .to_lowercase();
+                    if !haystack.contains(&text_filter) {
+                        continue;
+                    }
+                }
+
+                let json = serde_json::json!({
+                    "seq": entry.seq,
+                    "timestamp": entry.timestamp,
+                    "agent_id": entry.agent_id,
+                    "action": action_str,
+                    "detail": entry.detail,
+                    "outcome": entry.outcome,
+                    "hash": entry.hash,
+                });
+                let data = serde_json::to_string(&json).unwrap_or_default();
+                if tx.send(Ok(Event::default().data(data))).await.is_err() {
+                    return; // Client disconnected
+                }
+            }
+
+            // Update tracking state
+            if let Some(last) = entries.last() {
+                last_seq = last.seq;
+            }
+            first_poll = false;
+        }
+    });
+
+    let rx_stream = tokio_stream::wrappers::ReceiverStream::new(rx);
+    Sse::new(rx_stream)
+        .keep_alive(
+            KeepAlive::new()
+                .interval(std::time::Duration::from_secs(15))
+                .text("ping"),
+        )
+        .into_response()
+}
+
+/// Classify an audit action string into a level (info, warn, error).
+fn classify_audit_level(action: &str) -> &'static str {
+    let a = action.to_lowercase();
+    if a.contains("error") || a.contains("fail") || a.contains("crash") || a.contains("denied") {
+        "error"
+    } else if a.contains("warn") || a.contains("block") || a.contains("kill") {
+        "warn"
+    } else {
+        "info"
+    }
+}
