+++ crates/librefang-api/src/routes/mod.rs
@@ -20,11 +20,14 @@ pub mod audit;
 pub mod authz;
 pub mod auto_dream;
 pub mod backup;
+pub mod bindings;
 pub mod budget;
 pub mod channels;
+pub mod commands;
 pub mod config;
 pub mod goals;
 pub mod inbox;
+pub mod logs;
 pub mod mcp_auth;
 pub mod media;
 pub mod memory;
@@ -60,11 +63,14 @@ pub use audit::*;
 pub use authz::*;
 pub use auto_dream::*;
 pub use backup::*;
+pub use bindings::*;
 pub use budget::*;
 pub use channels::*;
+pub use commands::*;
 pub use config::*;
 pub use goals::*;
 pub use inbox::*;
+pub use logs::*;
 pub use mcp_auth::*;
 pub use media::*;
 pub use memory::*;
